Fix NumPy 2.x compatibility in trainer critic update
Cast critic_states slice to float32 before passing to model to fix
RuntimeError on NumPy 2.x where numpy.int32 dtype cannot be inferred
by torch.FloatTensor constructor in older PyTorch versions.

+++ satellite_mec/training/trainer.py
@@ -111,7 +111,7 @@ class MAPPOTrainer:
             # Critic 更新
             for start in range(0, n_slot, cfg.MINIBATCH):
                 idx    = slot_idx[start:min(start + cfg.MINIBATCH, n_slot)]
-                values = self.critic.get_value(critic_states[idx])
+                values = self.critic.get_value(critic_states[idx].astype(np.float32))
                 critic_loss = nn.MSELoss()(values, target_returns[idx])
                 self.critic_optimizer.zero_grad()
                 critic_loss.backward()
